wip (compiles)

+++ src/core/cost.rs
@@ -1,14 +1,14 @@
 #[derive(Debug, Clone, PartialEq, PartialOrd)]
 pub struct Cost {
     pub material_cost: u64,
-    pub leftover_value: f64,
+    pub leftover_value: f32,
     pub part_area_included: u64,
     pub part_area_excluded: u64,
 }
 
 
 impl Cost {
-    pub fn new(material_cost: u64, leftover_value: f64, part_area_included: u64, part_area_excluded: u64) -> Self {
+    pub fn new(material_cost: u64, leftover_value: f32, part_area_included: u64, part_area_excluded: u64) -> Self {
         Self { material_cost, leftover_value, part_area_included, part_area_excluded }
     }
